vector_pipeline/requirements.txt
boto3==1.39.7
PyPDF2==3.0.1
Pillow==11.2.1
PyMuPDF==1.26.3
pymupdf4llm==0.0.26
weaviate-client==4.15.4
pdf2image==1.17.0
PyYAML==6.0.2
torch==2.7.1
colpali_engine==0.3.11
